split reducer
