+++ src/store.js
@@ -1,4 +1,4 @@
 import { createStore } from 'redux';
-import { rootReducer } from './reducers';
+import rootReducer from './reducers/rootReducer';
 
 export const store = createStore(rootReducer)
